fix handler interfaces path
