+++ internal/server/auth/delivery/interface.go
@@ -1,0 +1,10 @@
+package delivery
+
+import "net/http"
+
+type AuthHandlerInterface interface {
+	Register(w http.ResponseWriter, r *http.Request)
+	Login(w http.ResponseWriter, r *http.Request)
+	Logout(w http.ResponseWriter, r *http.Request)
+	Session(w http.ResponseWriter, r *http.Request)
+}
