+++ internal/server/collection/delivery/interface.go
@@ -1,0 +1,7 @@
+package delivery
+
+import "net/http"
+
+type CollectionHandlerInterface interface {
+	GetMainPageCollections(w http.ResponseWriter, r *http.Request)
+}
