--- internal/server/router/api.go
@@ -1,22 +1,0 @@
-package router
-
-import "net/http"
-
-type AuthHandlerInterface interface {
-	Register(w http.ResponseWriter, r *http.Request)
-	Login(w http.ResponseWriter, r *http.Request)
-	Logout(w http.ResponseWriter, r *http.Request)
-	Session(w http.ResponseWriter, r *http.Request)
-}
-
-type StaffPersonHandlerInterface interface {
-	GetPerson(w http.ResponseWriter, r *http.Request)
-}
-
-type CollectionHandlerInterface interface {
-	GetMainPageCollections(w http.ResponseWriter, r *http.Request)
-}
-
-type UserHandlerInterface interface {
-	UpdateUser(w http.ResponseWriter, r *http.Request)
-}
