+++ internal/server/staff_person/delivery/interface.go
@@ -1,0 +1,7 @@
+package delivery
+
+import "net/http"
+
+type StaffPersonHandlerInterface interface {
+	GetPerson(w http.ResponseWriter, r *http.Request)
+}
