+++ internal/server/user/delivery/http/interface.go
@@ -1,0 +1,7 @@
+package http
+
+import "net/http"
+
+type UserHandlerInterface interface {
+	UpdateUser(w http.ResponseWriter, r *http.Request)
+}
